refactor: Move SupabaseClientWrapper to network package.
