+++ app/src/main/java/no/usn/mob3000/data/network/SupabaseClientWrapper.kt
@@ -1,4 +1,4 @@
-package no.usn.mob3000.data
+package no.usn.mob3000.data.network
 
 import android.content.Context
 import io.github.jan.supabase.SupabaseClient
@@ -8,6 +8,7 @@ import io.github.jan.supabase.gotrue.Auth
 import io.github.jan.supabase.postgrest.Postgrest
 import io.github.jan.supabase.realtime.Realtime
 import no.usn.mob3000.MainActivity
+import no.usn.mob3000.data.SecureEnvManager
 
 /**
  * Singleton object responsible for managing the Supabase client instance.
@@ -36,7 +37,8 @@ object SupabaseClientWrapper {
      */
     fun initialize(context: Context) {
         val supabaseUrl = SecureEnvManager.getDecryptedValue(context, "NEXT_PUBLIC_SUPABASE_URL")
-        val supabaseKey = SecureEnvManager.getDecryptedValue(context, "NEXT_PUBLIC_SUPABASE_ANON_KEY")
+        val supabaseKey =
+            SecureEnvManager.getDecryptedValue(context, "NEXT_PUBLIC_SUPABASE_ANON_KEY")
 
         if (supabaseUrl == null || supabaseKey == null) {
             throw IllegalStateException("Supabase credentials not found. Make sure to call SecureEnvManager.initializeEnvVariables(context) first.")
@@ -66,7 +68,7 @@ object SupabaseClientWrapper {
      * @created 2024-10-07
      */
     fun getClient(): SupabaseClient {
-        if (!::supabaseClient.isInitialized) {
+        if (!SupabaseClientWrapper::supabaseClient.isInitialized) {
             throw IllegalStateException("SupabaseClient not initialized. Call initialize(context) first.")
         }
