+++ app/src/main/java/no/usn/mob3000/ui/screens/chess/train/group/GroupsScreen.kt
@@ -24,7 +24,7 @@ import kotlinx.coroutines.withContext
 import kotlinx.serialization.Serializable
 import no.usn.mob3000.R
 import no.usn.mob3000.Viewport
-import no.usn.mob3000.data.SupabaseClientWrapper
+import no.usn.mob3000.data.network.SupabaseClientWrapper
 import no.usn.mob3000.ui.theme.DefaultButton
 import no.usn.mob3000.ui.screens.chess.train.opening.OpeningsScreen
 import no.usn.mob3000.ui.theme.DefaultListItemBackground
